+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.0.0-beta.3 - 2021-11-25
+
+### Added
+
+- `isPhe` attribute to Category client `TreeMenuItem` entity
+- `weeeFee` attribute to `Product` entity and `ProductRequest` DTO
+
 ## 4.0.0-beta.2 - 2021-11-22
 
 ### Added